exodt-backend/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.12.1
Automat==22.10.0
boto3==1.26.66
botocore==1.29.66
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
coreapi==2.3.3
coverage==7.0.1
cryptography==38.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.4
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-filter==22.1
django-jquery==3.1.0
django-oauth-toolkit==2.2.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
ecdsa==0.18.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jwcrypto==1.4.2
MarkupSafe==2.1.1
msgpack==1.0.4
oauthlib==3.2.2
Pillow==9.3.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
python-dateutil==2.8.2
python-decouple==3.6
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
wrapt==1.14.1
zope.interface==5.5.2
